\name{Antenatal care data}
\alias{ANCdata}
\docType{data}
\title{ Dataset on effect of new antenatal care method on mortality}
\description{
This dataset contains records of high risk pregnant women 
under a trial on new and old methods of antenatal care 
in two clinics. The outcome was perinatal mortality.
}
\usage{data(ANCdata)}
\format{
  A data frame with 755 observations on the following 3 variables.
  \describe{
    \item{\code{death}}{a factor with levels \code{no} \code{yes}}
    \item{\code{anc}}{a factor with levels \code{old} \code{new}}
    \item{\code{clinic}}{a factor with levels \code{A} \code{B}}
  }
}
\keyword{datasets}
