\name{Attitudes dataset}
\alias{Attitudes}
\docType{data}
\title{ Dataset from an attitude survey among hospital staff}
\description{
Survey on attitudes related to services among hospital staff. 

Codes for the answers qa1 to qa18 are
		\tabular{lll}{
        	\tab 1 \tab = strongly disagree\cr
         	\tab 2 \tab = disagree\cr
         	\tab 3 \tab = neutral\cr
         	\tab 4 \tab = agree\cr
         	\tab 5 \tab = strong agree\cr
       }

}
\usage{data(Attitudes)}
\format{
  A data frame with 136 observations on the following 7 variables.
  \describe{
    \item{\code{id}}{identifying code of repondent}
	\item{\code{sex}}{gender of respondent}
	\item{\code{dep}}{code of department}
	\item{\code{qa1}}{I have pride in my job}
	\item{\code{qa2}}{I'm happy to give service}
	\item{\code{qa3}}{I feel difficulty in giving service}
	\item{\code{qa4}}{I can improve my service}
	\item{\code{qa5}}{A service person must have patience}
	\item{\code{qa6}}{I would change my job if had the chance}
	\item{\code{qa7}}{Devoting some personal time will improve oneself}
	\item{\code{qa8}}{Hard work will improve oneself}
	\item{\code{qa9}}{Smiling leads to trust}
	\item{\code{qa10}}{I feel bad if I cannot give service}
	\item{\code{qa11}}{A client is not always right}
	\item{\code{qa12}}{Experienced clients should follow the procedure}
	\item{\code{qa13}}{A client violating the regulation should not bargain}
	\item{\code{qa14}}{Understanding colleagues will lead to understanding clients}
	\item{\code{qa15}}{Clients like this place due to good service}
	\item{\code{qa16}}{Clients who expect our smiling faces create pressure on us}
	\item{\code{qa17}}{Clients are often self-centered}
	\item{\code{qa18}}{Clients should be better served}
  }
}
\keyword{datasets}
