\name{Blood pressure}
\docType{data}
\alias{BP}
\title{Dataset on blood pressure and determinants}
\description{
This dataset contains information on the records of 100 adults from a small 
cross-sectional survey in 2001 investigating blood pressure and its determinants
in a community.
}
\usage{data(BP)}
\format{A data frame containing 100 observations and 6 variables with
variable descriptions.}
\examples{
data(BP)
des(BP)
}
\keyword{datasets}
    