\name{DHF99}
\alias{DHF99}
\docType{data}
\title{Dataset for exercise on predictors for mosquito larva infestation}
\description{
  Dataset from a community survey on water containers infested by mosquito larvae.
}
\usage{data(DHF99)}
\format{
  A data frame with 300 observations on the following 5 variables.
  \describe{
    \item{\code{houseid}}{a numeric vector}
    \item{\code{village}}{a numeric vector indicating village ID}
    \item{\code{education}}{a factor with levels \code{Primary} \code{Secondary} \code{High school} \code{Bachelor} \code{Other}}
    \item{\code{containers}}{a numeric vector indicating number of containers infested}
    \item{\code{viltype}}{a factor with levels \code{rural} \code{urban} \code{slum}}
  }
}
\references{
  Thammapalo, S., Chongsuwiwatwong, V., Geater, A., Lim, A., Choomalee, K. 2005. 
 Socio-demographic and environmental factors associated with Aedes breeding places in Phuket, Thailand.
\emph{Southeast Asian J Trop Med Pub Hlth} \bold{36(2)}: 426-33.}
\keyword{datasets}
