\name{IUD trial discontinuation data}
\alias{IudDiscontinue}
\docType{data}
\title{Dataset on discontinuation of the IUD trial cases}
\description{
This dataset is a subset of WHO IUD trial. It should be merged with IudAdmit and IudFollowup
}
\usage{data(IudDiscontinue)}
\format{A data frame containing 398 observations and 3 variables.
  \describe{
    \item{\code{id}}{a numeric vector for personal identification number}
    \item{\code{discdate}}{date of discontinuation}
    \item{\code{d23}}{primary reason for discontinuation}
  }
}
\examples{
data(IudDiscontinue)
}
\keyword{datasets}
    