\name{Montana}
\docType{data}
\alias{Montana}
\title{Dataset on arsenic exposure and respiratory deaths}
\description{
Dataset from a cohort study of exposure to arsenic from industry and deaths from respiratory diseases.
}
\format{
  A data frame with 114 observations on the following 6 variables.
  \describe{
    \item{\code{respdeath}}{a numeric vector indicating number of deaths from respiratory diseases}
    \item{\code{personyrs}}{a numeric vector indicating person-years of exposure}
    \item{\code{agegr}}{a numeric vector: 1=40-49, 2=50-59, 3=60-69, 4=70-79)}
    \item{\code{period}}{a numeric vector: 1=1938-1949, 2=1950-1959, 3=1960-1969, 4=1970-1977}
    \item{\code{starting}}{a numeric vector indicating starting period: 1=pre-1925, 2=1925 & after}
    \item{\code{arsenic}}{a numeric vector indicating years of exposure: 1=<1 year, 2=1-4 years, 3=5-14 years, 4=15+ years }
  }
}
\usage{data(Montana)}
\keyword{datasets}
    