\name{print kap.ByCategory}
\alias{print.kap.ByCategory}
\title{Print kap.ByCategory results}
\description{Print results for kap.Bycategory commands}
\usage{
\method{print}{kap.ByCategory}(x, ...)
}
\arguments{
       \item{x}{object of class 'kap.ByCategory'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'kap.ByCategory'}
\keyword{database}

