\name{print kap.table}
\alias{print.kap.table}
\title{Print kap.table results}
\description{Print results for kap.table commands}
\usage{
\method{print}{kap.table}(x, ...)
}
\arguments{
       \item{x}{object of class 'kap.table'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'kap.table'}
\keyword{database}

