\name{print matchTab}
\alias{print.matchTab}
\title{Print matched tabulation results}
\description{Print matched tabulation results}
\usage{
\method{print}{matchTab}(x, ...)
}
\arguments{
       \item{x}{object of class 'matchTab'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'matchTab'}
\keyword{database}                                                                            

