\name{print n.for.2p}
\alias{print.n.for.2p}
\title{Print n.for.2p results}
\description{Print results for sample size for hypothesis testing of 2 proportions}
\usage{
\method{print}{n.for.2p}(x, ...)
}
\arguments{
       \item{x}{object of class 'n.for.2p'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'n.for.2p'}
\examples{
n.for.2p(p1=.1, p2=.2)
n.for.2p(p1=seq(1,9,.5)/10, p2=.5)
}
\keyword{database}                                                                            

