\name{print n.for.cluster.2means}
\alias{print.n.for.cluster.2means}
\title{Print n.for.cluster.2means results}
\description{Print results for sample size for hypothesis testing of 2 means  in cluster RCT}
\usage{
\method{print}{n.for.cluster.2means}(x, ...)
}
\arguments{
       \item{x}{object of class 'n.for.cluster.2means'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'n.for.cluster.2means'}
\keyword{database}                                                                            

