\name{print n.for.cluster.2p}
\alias{print.n.for.cluster.2p}
\title{Print n.for.cluster.2p results}
\description{Print results for sample size for hypothesis testing of 2 proportions in cluster RCT}
\usage{
\method{print}{n.for.cluster.2p}(x, ...)
}
\arguments{
       \item{x}{object of class 'n.for.cluster.2p'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'n.for.cluster.2p'}
\keyword{database}                                                                            

