\name{print n.for.noninferior.2p}
\alias{print.n.for.noninferior.2p}
\title{Print n.for.noninferior.2p results}
\description{Print results for sample size for hypothesis testing of 2 proportions in non-inferior trial}
\usage{
\method{print}{n.for.noninferior.2p}(x, ...)
}
\arguments{
       \item{x}{object of class 'n.for.noninferior.2p'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'n.for.2p'}
\examples{
n.for.noninferior.2p(p=.85, sig.inferior=.05)
}
\keyword{database}                                                                            

