\name{print n.for.survey}
\alias{print.n.for.survey}
\title{Print n.for.survey results}
\description{Print results for sample size of a continuous variable}
\usage{
\method{print}{n.for.survey}(x, ...)
}
\arguments{
       \item{x}{object of class 'n.for.survey'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'n.for.2p'}
\examples{
n.for.survey(p=seq(5,95,5)/100)
}
\keyword{database}                                                                            

