\name{print statStack}
\alias{print.statStack}
\title{Print statStack object}
\description{Print a statStack object}
\usage{
\method{print}{statStack}(x, ...)
}
\arguments{
       \item{x}{object of class 'statStack' }
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'statStack'}
\keyword{database}

