\name{print summ.data.frame}
\alias{print.summ.data.frame}
\title{Print summary of the data frame}
\description{Print summary of data frame}
\usage{
\method{print}{summ.data.frame}(x, ...)
}
\arguments{
       \item{x}{object of class 'summ.data.frame'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'summ'}
\keyword{database}                                                                            

