\name{print summ.default}
\alias{print.summ.default}
\title{Print summary of a variable}
\description{Print summary of a variable}
\usage{
\method{print}{summ.default}(x, ...)
}
\arguments{
       \item{x}{object of class 'summ.default'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ cvirasak@gmail.com}
}
\seealso{'summ'}
\keyword{database}                                                                            

