\name{Voluntary counselling and testing}
\docType{data}
\alias{VCT}
\title{Dataset on attitudes toward VCT}
\description{
This dataset contains information on the records of 200 women working at a tourist destination community.
}
\usage{data(VCT)}
\format{A subset of a data frame containing 200 observations and 12 variables with
variable descriptions.

Details of the codes can be seen from the results of the function 'codebook()' below.
}
\examples{
data(VCT)
codebook(VCT)
}
\keyword{datasets}
    