\name{print des}
\alias{print.des}
\title{Print 'des' results}
\description{Print description of data frame of a variable}
\usage{
\method{print}{des}(x, ...)
}
\arguments{
       \item{x}{object of class 'des'}
       \item{...}{further arguments passed to or used by methods.}
}
\author{Virasakdi Chongsuvivatwong
       \email{ <cvirasak@medicine.psu.ac.th>}
}
\seealso{'des'}
\keyword{database}                                                                            

